creating poll working; created utils file
