+++ Clicker/Extensions/UIViewController+Shared.swift
@@ -12,29 +12,6 @@ import Alamofire
 import SwiftyJSON
 
 extension UIViewController {
-    // CONVERT INT TO MC OPTIONS
-    func intToMCOption(_ intOption: Int) -> String {
-        return String(Character(UnicodeScalar(intOption + Int(("A" as UnicodeScalar).value))!))
-    }
-    
-    // GET MM/DD/YYYY OF TODAY
-    func getTodaysDate() -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "MM/dd/yyyy"
-        return formatter.string(from: Date())
-    }
-    
-    // USER DEFAULTS
-    func encodeObjForKey(obj: Any, key: String) {
-        let encodedData = NSKeyedArchiver.archivedData(withRootObject: obj)
-        UserDefaults.standard.set(encodedData, forKey: key)
-    }
-    
-    func decodeObjForKey(key: String) -> Any {
-        let decodedData = UserDefaults.standard.value(forKey: key) as! Data
-        return NSKeyedUnarchiver.unarchiveObject(with: decodedData)!
-    }
-    
     func createAlert(title: String, message: String) -> UIAlertController {
         let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertControllerStyle.alert)
         alert.addAction(UIAlertAction(title: "Close", style: .cancel, handler: nil))
