+++ Clicker/View Controllers/BlackAnswerController.swift
@@ -134,7 +134,7 @@ class BlackAnswerController: UIViewController, UICollectionViewDelegate, UIColle
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cardRowCellIdentifier, for: indexPath) as! CardRowCell
         cell.polls = datePollsArr[indexPath.item].1
         cell.socket = socket
-        cell.pollRole = .ask
+        cell.pollRole = .answer
         return cell
     }
